refactor tests: split into individual test files per function
